Move files into appropriate directory structures.
